feat(elasticsearch): introduce new Elasticsearch client interface
This commit adds a new package for the Elasticsearch client interface, defining essential operations such as Ping, CreateIndex, and Search. The new structure enhances modularity and type safety, allowing for better integration with existing storage components. Additionally, references to the new internal client are updated across the codebase, ensuring consistency in client handling.
