+++ storage/elasticsearch/internal/elasticsearch/elasticsearch.go
@@ -7,32 +7,12 @@
 //
 //
 
-// Package elasticsearch provides Elasticsearch client interface.
+// Package elasticsearch provides Elasticsearch client interface and implementation.
 package elasticsearch
 
 import (
-	"context"
+	ielasticsearch "trpc.group/trpc-go/trpc-agent-go/internal/storage/elasticsearch"
 )
 
-// Client defines the minimal interface for Elasticsearch operations.
-// Use []byte payloads to decouple from SDK typed APIs.
-type Client interface {
-	// Ping checks if Elasticsearch is available.
-	Ping(ctx context.Context) error
-	// CreateIndex creates an index with the provided body.
-	CreateIndex(ctx context.Context, indexName string, body []byte) error
-	// DeleteIndex deletes the specified index.
-	DeleteIndex(ctx context.Context, indexName string) error
-	// IndexExists returns whether the specified index exists.
-	IndexExists(ctx context.Context, indexName string) (bool, error)
-	// IndexDoc indexes a document with the given identifier.
-	IndexDoc(ctx context.Context, indexName, id string, body []byte) error
-	// GetDoc retrieves a document by identifier and returns the raw body.
-	GetDoc(ctx context.Context, indexName, id string) ([]byte, error)
-	// UpdateDoc applies a partial update to the document by identifier.
-	UpdateDoc(ctx context.Context, indexName, id string, body []byte) error
-	// DeleteDoc deletes a document by identifier.
-	DeleteDoc(ctx context.Context, indexName, id string) error
-	// Search executes a query and returns the raw response body.
-	Search(ctx context.Context, indexName string, body []byte) ([]byte, error)
-}
+// Client is the Elasticsearch client interface.
+type Client = ielasticsearch.Client
